fix: shared element flicker of same named screen transitions
(check example)

In createSharedElementScene.tsx, changed  function to use route.key instead of route.name when checking to see if the current route is active. Previously, all screens that had the same name inthe stack think that they are active when the navigation event fires. This correctly finds the one screen that is active to perform the correct transition. In SharedElementRendererData, when updating share elements, use  variable to correctly send the screen's 'sharedElements' function the showing state. This fixes a bug where, if a user went backwards to a screen of the same name,  would be true when it should be false. Please take a look at the example I added. These fixes now allow for properly navigating back and forthwith a screen of the same name.